+++ crates/holochain/src/conductor/conductor.rs
@@ -39,6 +39,10 @@ use crate::conductor::config::ConductorConfig;
 use crate::conductor::error::ConductorResult;
 use crate::conductor::handle::ConductorHandle;
 use crate::core::queue_consumer::InitialQueueTriggers;
+use crate::core::ribosome::guest_callback::post_commit::PostCommitArgs;
+use crate::core::ribosome::guest_callback::post_commit::POST_COMMIT_CHANNEL_BOUND;
+use crate::core::ribosome::guest_callback::post_commit::POST_COMMIT_CONCURRENT_LIMIT;
+use crate::core::ribosome::RibosomeT;
 use crate::{
     conductor::api::error::ConductorApiResult, core::ribosome::real_ribosome::RealRibosome,
 };
@@ -52,6 +56,8 @@ use holochain_conductor_api::AppStatusFilter;
 use holochain_conductor_api::InstalledAppInfo;
 use holochain_conductor_api::IntegrationStateDump;
 use holochain_keystore::lair_keystore::spawn_lair_keystore;
+use holochain_keystore::lair_keystore::spawn_new_lair_keystore;
+use holochain_keystore::test_keystore::spawn_legacy_test_keystore;
 use holochain_keystore::test_keystore::spawn_test_keystore;
 use holochain_keystore::MetaLairClient;
 use holochain_sqlite::db::DbKind;
@@ -65,6 +71,7 @@ use std::collections::{HashMap, HashSet};
 use std::path::PathBuf;
 use std::sync::atomic::AtomicBool;
 use std::sync::Arc;
+use tokio::sync::mpsc::error::SendError;
 use tracing::*;
 
 #[cfg(any(test, feature = "test_utils"))]
@@ -166,6 +173,8 @@ where
 
     /// Database sync level
     db_sync_level: DbSyncLevel,
+
+    post_commit: tokio::sync::mpsc::Sender<PostCommitArgs>,
 }
 
 impl Conductor {
@@ -580,14 +589,15 @@ where
 
         // Clean up all cells that will be dropped (leave network, etc.)
         let to_cleanup: Vec<_> = self.cells.share_mut(|cells| {
-            let to_cleanup: Vec<_> = cells
+            let to_remove = cells
                 .keys()
                 .filter(|id| !keepers.contains(id))
                 .cloned()
-                .collect();
-            to_cleanup
-                .into_iter()
-                .filter_map(|cell_id| cells.remove(&cell_id))
+                .collect::<Vec<_>>();
+
+            to_remove
+                .iter()
+                .filter_map(|cell_id| cells.remove(cell_id))
                 .collect()
         });
         for cell in to_cleanup {
@@ -834,7 +844,7 @@ where
                         dna_def
                             .zomes
                             .iter()
-                            .map(|(zome_name, zome)| Ok(zome.wasm_hash(&zome_name)?))
+                            .map(|(zome_name, zome)| Ok(zome.wasm_hash(zome_name)?))
                     })
                     .collect::<ConductorResult<HashSet<_>>>()?;
 
@@ -1064,6 +1074,13 @@ where
             Ok(())
         })
     }
+
+    /// Get the post commit sender.
+    pub async fn post_commit_permit(
+        &self,
+    ) -> Result<tokio::sync::mpsc::OwnedPermit<PostCommitArgs>, SendError<()>> {
+        self.post_commit.clone().reserve_owned().await
+    }
 }
 
 /// Perform Genesis on the source chains for each of the specified CellIds.
@@ -1186,6 +1203,7 @@ where
         root_env_dir: EnvironmentRootPath,
         holochain_p2p: holochain_p2p::HolochainP2pRef,
         db_sync_level: DbSyncLevel,
+        post_commit: tokio::sync::mpsc::Sender<PostCommitArgs>,
     ) -> ConductorResult<Self> {
         Ok(Self {
             conductor_env: env,
@@ -1201,6 +1219,7 @@ where
             root_env_dir,
             holochain_p2p,
             db_sync_level,
+            post_commit,
         })
     }
 
@@ -1292,6 +1311,8 @@ mod builder {
         pub config: ConductorConfig,
         /// The DnaStore (mockable)
         pub dna_store: DS,
+        /// For new lair, passphrase is required
+        pub passphrase: Option<sodoken::BufRead>,
         /// Optional keystore override
         pub keystore: Option<MetaLairClient>,
         #[cfg(any(test, feature = "test_utils"))]
@@ -1329,6 +1350,12 @@ mod builder {
             self
         }
 
+        /// Set the passphrase for use in keystore initialization
+        pub fn passphrase(mut self, passphrase: Option<sodoken::BufRead>) -> Self {
+            self.passphrase = passphrase;
+            self
+        }
+
         /// Initialize a "production" Conductor
         pub async fn build(self) -> ConductorResult<ConductorHandle> {
             cfg_if::cfg_if! {
@@ -1349,11 +1376,7 @@ mod builder {
                 match &self.config.keystore {
                     KeystoreConfig::DangerTestKeystoreLegacyDeprecated => {
                         tracing::warn!("Using DEPRECATED legacy lair api.");
-                        let keystore = spawn_test_keystore().await?;
-                        // pre-populate with our two fixture agent keypairs
-                        keystore.new_sign_keypair_random().await.unwrap();
-                        keystore.new_sign_keypair_random().await.unwrap();
-                        keystore
+                        spawn_legacy_test_keystore().await?
                     }
                     KeystoreConfig::LairServerLegacyDeprecated {
                         keystore_path,
@@ -1365,6 +1388,19 @@ mod builder {
                             danger_passphrase_insecure_from_config.as_bytes(),
                         );
                         spawn_lair_keystore(keystore_path.as_deref(), passphrase).await?
+                    }
+                    KeystoreConfig::DangerTestKeystore => spawn_test_keystore().await?,
+                    KeystoreConfig::LairServer { connection_url } => {
+                        let passphrase = match self.passphrase {
+                            None => {
+                                return Err(one_err::OneErr::new(
+                                    "passphrase required for new lair keystore api",
+                                )
+                                .into())
+                            }
+                            Some(p) => p,
+                        };
+                        spawn_new_lair_keystore(connection_url.clone(), passphrase).await?
                     }
                     oth => unimplemented!("unimplemented keystore config: {:?}", oth),
                 }
@@ -1408,6 +1444,9 @@ mod builder {
             let (holochain_p2p, p2p_evt) =
                 holochain_p2p::spawn_holochain_p2p(network_config, tls_config).await?;
 
+            let (post_commit_sender, post_commit_receiver) =
+                tokio::sync::mpsc::channel(POST_COMMIT_CHANNEL_BOUND);
+
             let conductor = Conductor::new(
                 environment,
                 wasm_environment,
@@ -1416,6 +1455,7 @@ mod builder {
                 env_path,
                 holochain_p2p,
                 config.db_sync_level,
+                post_commit_sender,
             )
             .await?;
 
@@ -1441,19 +1481,60 @@ mod builder {
                 dev_settings: parking_lot::RwLock::new(DevSettings::default()),
             });
 
-            Self::finish(handle, config, p2p_evt).await
+            Self::finish(handle, config, p2p_evt, post_commit_receiver).await
+        }
+
+        fn spawn_post_commit(
+            conductor_handle: ConductorHandle,
+            receiver: tokio::sync::mpsc::Receiver<PostCommitArgs>,
+        ) {
+            let receiver_stream = tokio_stream::wrappers::ReceiverStream::new(receiver);
+            tokio::task::spawn(receiver_stream.for_each_concurrent(
+                POST_COMMIT_CONCURRENT_LIMIT,
+                move |post_commit_args| {
+                    let conductor_handle = conductor_handle.clone();
+                    async move {
+                        let PostCommitArgs {
+                            host_access,
+                            invocation,
+                            cell_id,
+                        } = post_commit_args;
+                        match conductor_handle.clone().get_ribosome(cell_id.dna_hash()) {
+                            Ok(ribosome) => {
+                                if let Err(e) = tokio::task::spawn_blocking(move || {
+                                    if let Err(e) =
+                                        ribosome.run_post_commit(host_access, invocation)
+                                    {
+                                        tracing::error!(?e);
+                                    }
+                                })
+                                .await
+                                {
+                                    tracing::error!(?e);
+                                }
+                            }
+                            Err(e) => {
+                                tracing::error!(?e);
+                            }
+                        }
+                    }
+                },
+            ));
         }
 
         async fn finish(
             handle: ConductorHandle,
             conductor_config: ConductorConfig,
             p2p_evt: holochain_p2p::event::HolochainP2pEventReceiver,
+            post_commit_receiver: tokio::sync::mpsc::Receiver<PostCommitArgs>,
         ) -> ConductorResult<ConductorHandle> {
             tokio::task::spawn(p2p_event_task(p2p_evt, handle.clone()));
 
             let _ = handle
                 .clone()
                 .start_scheduler(holochain_zome_types::schedule::SCHEDULER_INTERVAL);
+
+            let _ = Self::spawn_post_commit(handle.clone(), post_commit_receiver);
 
             let configs = conductor_config.admin_interfaces.unwrap_or_default();
             let cell_startup_errors = handle.clone().initialize_conductor(configs).await?;
@@ -1519,6 +1600,9 @@ mod builder {
                 holochain_p2p::spawn_holochain_p2p(self.config.network.clone().unwrap_or_default(), holochain_p2p::kitsune_p2p::dependencies::kitsune_p2p_proxy::TlsConfig::new_ephemeral().await.unwrap())
                     .await?;
 
+            let (post_commit_sender, post_commit_receiver) =
+                tokio::sync::mpsc::channel(POST_COMMIT_CHANNEL_BOUND);
+
             let conductor = Conductor::new(
                 envs.conductor(),
                 envs.wasm(),
@@ -1527,6 +1611,7 @@ mod builder {
                 self.config.environment_path.clone(),
                 holochain_p2p,
                 self.config.db_sync_level,
+                post_commit_sender,
             )
             .await?;
 
@@ -1560,7 +1645,7 @@ mod builder {
                     .expect("Could not install DNA");
             }
 
-            Self::finish(handle, self.config, p2p_evt).await
+            Self::finish(handle, self.config, p2p_evt, post_commit_receiver).await
         }
     }
 }