+++ crates/holochain/CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - adds new lair as an additional dependency and begins preparation for a config-time switch allowing use of new api lair keystore.
 - Add method `SweetDnaFile::from_bundle_with_overrides` [\#1030](https://github.com/holochain/holochain/pull/1030)
 - Some `SweetConductor::setup_app_*` methods now take anything iterable, instead of array slices, for specifying lists of agents and DNAs [\#1030](https://github.com/holochain/holochain/pull/1030)
+- `post_commit` hook is implemented now [PR 1000](https://github.com/holochain/holochain/pull/1000)
 - BREAKING conductor config changes [\#1031](https://github.com/holochain/holochain/pull/1031)
 
 Where previously, you might have had:
@@ -41,6 +42,21 @@ keystore:
   type: danger_test_keystore_legacy_deprecated
 ```
 - Bump legacy lair version to 0.0.8 fixing a crash when error message was too long [#1046](https://github.com/holochain/holochain/pull/1046)
+
+- Options to use new lair keystore [#1040](https://github.com/holochain/holochain/pull/1040)
+
+```yaml
+keystore:
+  type: danger_test_keystore
+```
+
+or
+
+```yaml
+keystore:
+  type: lair_server
+  connection_url: "unix:///my/path/socket?k=Foo"
+```
 
 ## 0.0.109